// src/types/Property/PropertyDashboardDTO.ts
export interface PropertyDashboardDTO {
    id: string;
    title: string;
    description: string;
    location: string;
    city: string;
    state: string;
    country: string;
    zip: string;
    createdAt: string;
    lastUpdatedAt: string;
    approvalStatus: string;
    approvalComment: string;
    imageUrl: string;
    price: number;
    priceLabel: string;
}